requirements.txt
anytree
aria2p
asyncio
beautifulsoup4
bencoding
cfscrape
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
tenacity
tgCrypto
xattr
yt-dlp
bs4
cloudscraper==1.2.58
dnspython
gdown
js2py
markdown
mutagen
psutil
pymongo
pySmartDL
telegraph
validators
megatools
